+++ src/BootstrapTable.js
@@ -59,6 +59,10 @@ class BootstrapTable extends Component {
 
     const isKeyFieldDefined = typeof keyField === 'string' && keyField.length;
     React.Children.forEach(props.children, column => {
+      if (column === null || column === undefined) {
+        // Skip null and undefined value
+        return;
+      }
       if (column.props.isKey) {
         if (keyField) {
           throw new Error('Error. Multiple key column be detected in TableHeaderColumn.');
@@ -142,11 +146,21 @@ class BootstrapTable extends Component {
   getColumnsDescription({ children }) {
     let rowCount = 0;
     React.Children.forEach(children, (column) => {
+      if (column === null || column === undefined) {
+        // Skip null and undefined value
+        return;
+      }
+
       if (Number(column.props.row) > rowCount) {
         rowCount = Number(column.props.row);
       }
     });
     return React.Children.map(children, (column, i) => {
+      if (column === null || column === undefined) {
+        // Return null for empty objects
+        return null;
+      }
+
       const rowIndex = column.props.row ? Number(column.props.row) : 0;
       const rowSpan = column.props.rowSpan ? Number(column.props.rowSpan) : 1;
       if ((rowSpan + rowIndex) === (rowCount + 1)) {
@@ -284,7 +298,9 @@ class BootstrapTable extends Component {
 
   componentWillUnmount() {
     window.removeEventListener('resize', this._adjustTable);
-    this.refs.body.refs.container.removeEventListener('scroll', this._scrollHeader);
+    if (this.refs && this.refs.body && this.refs.body.refs) {
+      this.refs.body.refs.container.removeEventListener('scroll', this._scrollHeader);
+    }
     if (this.filter) {
       this.filter.removeAllListeners('onFilterChange');
     }
@@ -503,7 +519,11 @@ class BootstrapTable extends Component {
     });
   }
 
-  handleExpandRow = expanding => {
+  handleExpandRow = (expanding, rowKey, isRowExpanding) => {
+    const { onExpand } = this.props.options;
+    if (onExpand) {
+      onExpand(rowKey, !isRowExpanding);
+    }
     this.setState({ expanding, reset: false }, () => {
       this._adjustHeaderWidth();
     });
@@ -966,7 +986,7 @@ class BootstrapTable extends Component {
     }
 
     const keys = [];
-    this.props.children.map(function(column) {
+    this.props.children.filter(_ => _ != null).map(function(column) {
       if (column.props.export === true ||
         (typeof column.props.export === 'undefined' &&
         column.props.hidden === false)) {
@@ -1101,7 +1121,8 @@ class BootstrapTable extends Component {
       || this.props.options.toolBar) {
       let columns;
       if (Array.isArray(children)) {
-        columns = children.map((column, r) => {
+        columns = children.filter(_ => _ != null).map((column, r) => {
+          if (!column) return;
           const { props } = column;
           const isKey = props.isKey || keyField === props.dataField;
           return {
@@ -1253,7 +1274,7 @@ class BootstrapTable extends Component {
         }
       }
     } else {
-      React.Children.forEach(this.props.children, (child, i) => {
+      React.Children.forEach(this.props.children.filter(_ => !!_), (child, i) => {
         if (child.props.width) {
           header[i].style.width = `${child.props.width}px`;
           header[i].style.minWidth = `${child.props.width}px`;
@@ -1454,6 +1475,7 @@ BootstrapTable.propTypes = {
     expandRowBgColor: PropTypes.string,
     expandBy: PropTypes.string,
     expanding: PropTypes.array,
+    onExpand: PropTypes.func,
     onlyOneExpanding: PropTypes.bool,
     beforeShowError: PropTypes.func,
     printToolBar: PropTypes.bool,
@@ -1602,6 +1624,7 @@ BootstrapTable.defaultProps = {
     expandRowBgColor: undefined,
     expandBy: Const.EXPAND_BY_ROW,
     expanding: [],
+    onExpand: undefined,
     onlyOneExpanding: false,
     beforeShowError: undefined,
     printToolBar: true,