+++ src/ExpandComponent.js
@@ -1,18 +1,17 @@
 /* eslint max-len: 0 */
+/* eslint no-nested-ternary: 0 */
 import React, { Component } from 'react';
 import classSet from 'classnames';
 
 class ExpandComponent extends Component {
 
   render() {
+    const { className } = this.props;
     const trCss = {
       style: {
         backgroundColor: this.props.bgColor
       },
-      className: classSet(
-        this.props.isSelected ? this.props.selectRow.className : null,
-        this.props.className
-      )
+      className: classSet(className)
     };
     return (
       <tr hidden={ this.props.hidden } width={ this.props.width } { ...trCss }>